🧪 refactor(isDomainAllowed): change directory, add tests (#2539)
