+++ api/server/services/isDomainAllowed.js
@@ -1,0 +1,24 @@
+const getCustomConfig = require('~/server/services/Config/getCustomConfig');
+
+async function isDomainAllowed(email) {
+  if (!email) {
+    return false;
+  }
+
+  const domain = email.split('@')[1];
+
+  if (!domain) {
+    return false;
+  }
+
+  const customConfig = await getCustomConfig();
+  if (!customConfig) {
+    return true;
+  } else if (!customConfig?.registration?.allowedDomains) {
+    return true;
+  }
+
+  return customConfig.registration.allowedDomains.includes(domain);
+}
+
+module.exports = isDomainAllowed;
